20-11 updated script for PanelApp evidence.
